Phase 2 step 14: Migrate LifecycleMixin + EventsActionsMixin LAST (state-birth site + dispatch surface)
